# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Enterprise Intelligence AI Platform API",
    "version": "1.0.0",
    "description": "API documentation for the Enterprise Intelligence AI Transformation Platform Hub (EIAPH)",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "API Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "sessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "corevia.sid",
        "description": "Session-based authentication"
      }
    },
    "schemas": {
...
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {},
          "message": {
            "type": "string"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "example": "Error message"
          },
          "status": {
            "type": "integer",
            "example": 400
          },
          "details": {}
        }
      },
...
      "CreateDemandVersionRequest": {
        "type": "object",
        "properties": {
          "versionType": {
            "type": "string",
            "enum": [
              "major",
              "minor",
              "patch"
            ]
          },
          "contentType": {
            "type": "string",
            "enum": [
              "business_case",
              "requirements",
              "enterprise_architecture",
              "strategic_fit"
            ],
            "default": "business_case"
          },
          "changesSummary": {
            "type": "string"
          },
          "skipAiSummary": {
            "type": "boolean",
            "default": false
          },
          "editReason": {
            "type": "string"
          },
          "createdBy": {
            "type": "string"
          },
          "createdByName": {
            "type": "string"
          },
          "createdByRole": {
            "type": "string"
          },
          "createdByDepartment": {
            "type": "string"
          },
          "sessionId": {
            "type": "string"
          },
          "ipAddress": {
            "type": "string"
          },
          "businessCaseId": {
            "type": "string"
          },
          "editedContent": {
            "type": "object",
            "additionalProperties": true
          },
          "teamAssignments": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": true
            }
          }
        },
        "required": [
          "versionType",
          "createdBy",
          "createdByName"
        ]
      },
      "UpdateDemandVersionRequest": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "under_review",
              "approved",
              "published",
              "archived",
              "rejected",
              "superseded"
            ]
          },
          "reviewComments": {
            "type": "string"
          },
          "approvalComments": {
            "type": "string"
          },
          "updatedBy": {
            "type": "string"
          },
          "updatedByName": {
            "type": "string"
          },
          "updatedByRole": {
            "type": "string"
          },
          "sessionId": {
            "type": "string"
          },
          "ipAddress": {
            "type": "string"
          }
        },
        "required": [
          "updatedBy",
          "updatedByName"
        ]
      },
      "ApproveDemandVersionRequest": {
        "type": "object",
        "properties": {
          "approvalComments": {
            "type": "string"
          },
          "approvalLevel": {
            "type": "string",
            "enum": [
              "initial",
              "manager",
              "final"
            ],
            "default": "initial"
          },
          "sessionId": {
            "type": "string"
          },
          "ipAddress": {
            "type": "string"
          }
        }
      },
...
      "DemandVersionListResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SuccessResponse"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "array",
                "items": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          }
        ]
      },
...
    }
  },
...
  "paths": {
...
    "/demand-reports/{id}/versions": {
      "post": {
        "tags": [
          "Demand Reports"
        ],
        "summary": "Create a report version",
        "operationId": "createDemandReportVersion",
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDemandVersionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Version created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation failure or pending version conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Demand report not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Demand Reports"
        ],
        "summary": "List report versions",
        "operationId": "listDemandReportVersions",
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Version list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DemandVersionListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Demand report not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/demand-reports/{id}/versions/{versionId}": {
      "get": {
        "tags": [
          "Demand Reports"
        ],
        "summary": "Get a specific report version",
        "operationId": "getDemandReportVersion",
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "versionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Version detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Version not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Demand Reports"
        ],
        "summary": "Update a specific report version",
        "operationId": "updateDemandReportVersion",
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "versionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDemandVersionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Version updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Version not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/demand-reports/{id}/versions/{versionId}/approve": {
      "post": {
        "tags": [
          "Demand Reports"
        ],
        "summary": "Approve a report version",
        "operationId": "approveDemandReportVersion",
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "versionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveDemandVersionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Version approved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation or workflow failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Version not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}